style: remove home-rolled `max` and `abs` macros
This prevents Clippy from triggering the `suboptimal-flops` and
`cast-lossless` lints. The casting up to `f64` should be cheap, means
we can use the built-in float methods, and once `f128` is eventually
stabilised, this crate will return a compile error instead of silently
returning incorrect results for `f128` values.

It also looked like `afe_max` was just entirely unused and intentionally
hidden from the docs, so I've deleted it.